requirements.txt
# --- Core ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Database ---
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2

# --- Pydantic & settings ---
pydantic==2.9.2
pydantic-settings==2.4.0
email-validator==2.2.0

# --- Auth / Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9

# --- Utils ---
python-dotenv==1.0.1
